backend/pyproject.toml
[tool.poetry]
name = "drk"
version = "0.1.0"
description = "PoC of Django + React integration with KeyCloack"
authors = ["AlTosterino"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.0.0"
requests = "^2.31.0"
pyjwt = "^2.7.0"
cryptography = "^41.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.270"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
